PRMT-4566 reorganise code
