+++ src/services/database/ehr-core-repository.js
@@ -1,9 +1,8 @@
-import { EhrTransferTracker } from './dynamo-ehr-transfer-tracker';
-import { arrayOfFragments, isFragment } from "../../models/fragment";
-import { core, isCore } from "../../models/core";
-import { RecordType } from '../../models/enums';
-import { logError } from '../../middleware/logging';
-import { MessageNotFoundError } from "../../errors/errors";
+import { EhrTransferTracker } from "./dynamo-ehr-transfer-tracker";
+import { arrayOfFragments } from "../../models/fragment";
+import { core } from "../../models/core";
+import { RecordType } from "../../models/enums";
+import { logError } from "../../middleware/logging";
 
 export const createCore = async ({ conversationId, messageId, fragmentMessageIds }) => {
   // to replace the existing `createEhrExtract` method
@@ -33,20 +32,3 @@ export const getCoreByKey = (inboundConversationId, inboundMessageId) => {
   return db.getItemByKey(inboundConversationId, inboundMessageId, RecordType.CORE);
 };
 
-export const getHealthRecordMessageIds = async (conversationId) => {
-  // to replace the method of same name
-
-  const db = EhrTransferTracker.getInstance();
-  const items = await db.queryTableByConversationId(conversationId, RecordType.ALL);
-
-  const core = items.filter(isCore)?.[0];
-  const fragments = items.filter(isFragment);
-
-  if (!core) {
-    throw new MessageNotFoundError();
-  }
-  const coreMessageId = core.InboundMessageId;
-  const fragmentMessageIds = fragments.map(message => message.InboundMessageId);
-
-  return { coreMessageId, fragmentMessageIds };
-}
